+++ internal/north/rest/hub_test.go
@@ -43,6 +43,8 @@ func (f *fakeHubIndex) SerialSuffix(central string) string {
 	return ""
 }
 
+func (f *fakeHubIndex) Released(string) bool { return true }
+
 type fakeProgramWriter struct {
 	calls atomic.Int32
 	id    atomic.Value